# module/vmanager/driver/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "certus VNFM Driver Service rest API"
  },
  "basePath": "/api/{vnfmtype}/v1",
...
  "paths": {
    "/{vnfmid}/vnfs": {
      "post": {
        "tags": [
          "vnf instantiate"
        ],
        "summary": "instantiate the vnf",
        "description": "",
        "operationId": "instantiate_vnf",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "vnfmid",
            "in": "path",
            "description": "vnfm instance id",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "request parameters",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VnfRequestParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/VnfResult"
            }
          },
          "404": {
            "description": "the vnfm instance id is wrong"
          },
          "500": {
            "description": "the url is invalid"
          }
        }
      }
    },
...
    "/{vnfmid}/vnfs/{vnfInstanceId}": {
      "get": {
        "tags": [
          "query vnf"
        ],
        "summary": "query the vnf",
        "description": "",
        "operationId": "query_vnf",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "vnfmid",
            "in": "path",
            "description": "vnfm instance id",
            "required": true,
            "type": "string"
          },
          {
            "name": "vnfInstanceId",
            "in": "path",
            "description": "vnf instance id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/returnVnfInfo"
            }
          },
          "404": {
            "description": "the vnfmid and vnfInstanceId  are wrong"
          },
          "500": {
            "description": "the url is invalid"
          }
        }
      }
    },
...
  },
  "definitions": {
    "VnfRequestParams": {
      "type": "object",
      "properties": {
        "vnfInstanceName": {
          "type": "string"
        },
        "vnfPackageId": {
          "type": "string"
        },
        "vnfDescriptorId": {
          "type": "string"
        },
        "vnfInstanceDescription": {
          "type": "string"
        },
        "additionalParam": {
          "type": "object",
          "properties": {
            "sdncontroller": {
              "type": "string"
            },
            "NatIpRange": {
              "type": "string"
            },
            "m6000_mng_ip": {
              "type": "string"
            },
            "externalPluginManageNetworkName": {
              "type": "string"
            },
            "location": {
              "type": "string"
            },
            "externalManageNetworkName": {
              "type": "string"
            },
            "sfc_data_network": {
              "type": "string"
            },
            "externalDataNetworkName": {
              "type": "string"
            },
            "inputs": {
              "type": "object"
            }
          }
        }
      }
    },
...
    "VnfResult": {
      "type": "object",
      "properties": {
        "vnfInstanceId": {
          "type": "string"
        },
        "jobid": {
          "type": "string"
        }
      }
    },
    "returnVnfInfo": {
      "type": "object",
      "properties": {
        "vnfInfo": {
          "type": "object",
          "properties": {
            "nfInstanceId": {
              "type": "string"
            },
            "vnfStatus": {
              "type": "string"
            },
            "version": {
              "type": "string"
            }
          }
        }
      }
    },
...
  }
}